Changed Short(short) to Integer(int) for status code management

+++ keyple-core/src/main/java/org/eclipse/keyple/seproxy/ApduResponse.java
@@ -48,7 +48,7 @@ public final class ApduResponse implements Serializable {
      * @param buffer apdu response data buffer (including sw1sw2)
      * @param successfulStatusCodes optional list of successful status codes other than 0x9000
      */
-    public ApduResponse(byte[] buffer, Set<Short> successfulStatusCodes) {
+    public ApduResponse(byte[] buffer, Set<Integer> successfulStatusCodes) {
 
         this.bytes = buffer;
         if (buffer == null) {
@@ -65,7 +65,7 @@ public final class ApduResponse implements Serializable {
             }
             if (successfulStatusCodes != null) {
                 this.successful =
-                        statusCode == 0x9000 || successfulStatusCodes.contains((short) statusCode);
+                        statusCode == 0x9000 || successfulStatusCodes.contains(statusCode);
             } else {
                 this.successful = statusCode == 0x9000;
             }
